==> 03_External_Validation/01_Create_sets.ipynb <==
import csv
import pandas as pd
import numpy as np

# mimic remove mean and special characters from dataset

# Read MIMIC the CSV file
mimic_df = pd.read_csv('..\\Collect_Chart_Lab_Events\\CSV\\Exports\\datasets\\whole_set\\o05_final_mean_table_los.csv')

# Read eICU the CSV file
eicu_df = pd.read_csv('..\\Collect_data_eicu\\CSV\\Exports\\datasets\\whole_set\\o01_final_mean_table.csv')

"""Summarize GCS and Braden"""

# Sum the GSC columns and if the value is NaN then we excluded it from the summation.
mimic_df['GCS  - Mean'] = mimic_df.apply(lambda row: row[['GCS - Eye Opening  - Mean', 'GCS - Verbal Response  - Mean', 'GCS - Motor Response  - Mean']].sum() if not all(row[['GCS - Eye Opening  - Mean', 'GCS - Verbal Response  - Mean', 'GCS - Motor Response  - Mean']].isna()) else np.nan, axis=1)

# Sum the Braden columns and if the value is NaN then we excluded it from the summation.
mimic_df['Braden  - Mean'] = mimic_df.apply(lambda row: row[['Braden Sensory Perception  - Mean', 'Braden Moisture  - Mean', 'Braden Activity  - Mean', 'Braden Mobility  - Mean', 'Braden Nutrition  - Mean', 'Braden Friction/Shear  - Mean']].sum() if not all(row[['Braden Sensory Perception  - Mean', 'Braden Moisture  - Mean', 'Braden Activity  - Mean', 'Braden Mobility  - Mean', 'Braden Nutrition  - Mean', 'Braden Friction/Shear  - Mean']].isna()) else np.nan, axis=1)

"Mimic remove mean and spaces"

# Remove "Mean" from column headers
mimic_df.columns = [col.replace(' - Mean', '') if ' - Mean' in col else col for col in mimic_df.columns]

# Remove "," from column headers
mimic_df.columns = [col.replace(',', '') if ',' in col else col for col in mimic_df.columns]

# remove blank spaces from the end of header names
mimic_df.columns = mimic_df.columns.str.strip()

"""-------------------------------------------------------------------------------------------------------"""

"""eICU remove mean and spaces"""
# Remove "Mean" from column headers
eicu_df.columns = [col.replace('(Mean)', '') if '(Mean)' in col else col for col in eicu_df.columns]

# Remove "," from column headers
eicu_df.columns = [col.replace(',', '') if ',' in col else col for col in eicu_df.columns]

# remove blank spaces from the end of header names
eicu_df.columns = eicu_df.columns.str.strip()

# Mimic -> columns to keep
# Read the CSV file to get the columns to keep
columns_to_keep = pd.read_csv('CSV\\imports\\mimic_features.csv')

# Drop second column from the column_names_df
columns_to_keep.drop(columns=['Unnamed: 1'], inplace=True)

# Extract column names from columns_to_keep DataFrame
columns_to_keep_names = columns_to_keep['column'].tolist()

# Select only the desired columns
mimic_temp = mimic_df[columns_to_keep_names]

"""------------------------------------------------------------------------------"""
# eICU -> columns to keep
# Read the CSV file to get the columns to keep
columns_to_keep = pd.read_csv("CSV\\imports\\eicu_features.csv")

# Drop second column from the column_names_df
columns_to_keep.drop(columns=['Unnamed: 1'], inplace=True)

# Extract column names from columns_to_keep DataFrame
columns_to_keep_names = columns_to_keep['column'].tolist()

# Select only the desired columns
eicu_temp = eicu_df[columns_to_keep_names]

# eICU -> columns to keep

# Read the CSV file to get the columns to keep
columns_to_keep = pd.read_csv("CSV\\imports\\eicu_features.csv")

# Drop second column from the column_names_df
columns_to_keep.drop(columns=['Unnamed: 1'], inplace=True)

# Extract column names from columns_to_keep DataFrame
columns_to_keep_names = columns_to_keep['column'].tolist()

# Select only the desired columns
eicu_temp = eicu_df[columns_to_keep_names]

"""--------Replace Block----------"""


# Replace 'Alive' with 0 and 'Expired' with 1 in the 'unitdischargestatus' column
eicu_temp['unitdischargestatus'] = eicu_temp['unitdischargestatus'].replace({'Alive': 0, 'Expired': 1})

# Replace 'Alive' with 0 and 'Expired' with 1 in the 'unitdischargestatus' column
eicu_temp['gender'] = eicu_temp['gender'].replace({'Female': 'F', 'Male': 'M'})

# Multiply values by 4, leaving NaN values unchanged
mimic_temp['Ionized Calcium'] = mimic_temp['Ionized Calcium'].apply(lambda x: x * 4 if not pd.isna(x) else x)

# Replace values in the 'ethnicity' column
eicu_temp['ethnicity'] = eicu_temp['ethnicity'].replace({
    'African American': 'BLACK/AFRICAN AMERICAN',
    'Caucasian': 'WHITE',
    'Hispanic': 'HISPANIC OR LATINO',
    'Asian': 'ASIAN',
    'Native American': 'AMERICAN INDIAN/ALASKA NATIVE',
    'Other/Unknown': 'UNKNOWN'
})

# Replace age from eicu which are higher that 89
eicu_temp['age'] = eicu_temp['age'].replace('> 89', '90')
eicu_temp['age'] = eicu_temp['age'].astype(np.int64)

column_eicu_mapping = {
    'column': 'column',
    'row_count': 'row_count',
    'uniquepid': 'subject_id',
    'patientunitstayid': 'hadm_id',
    'Time_Zone': 'Time_Zone',
    'gender': 'gender',
    'age': 'age',
    'ethnicity': 'race',
    'Base Excess': 'Base Excess',
    'lactate': 'Lactate',
    'paCO2': 'pCO2',
    'Total CO2': 'Calculated Total CO2',
    'BUN': 'BUN',
    'pH': 'pH',
    'paO2': 'pO2',
    'ALT (SGPT)': 'Alanine Aminotransferase (ALT)',
    'alkaline phos.': 'Alkaline Phosphatase',
    'anion gap': 'Anion Gap',
    'AST (SGOT)': 'Asparate Aminotransferase (AST)',
    'bicarbonate': 'Bicarbonate',
    'chloride': 'Chloride',
    'creatinine': 'Creatinine',
    'glucose': 'Glucose',
    'magnesium': 'Magnesium',
    'phosphate': 'Phosphate',
    'potassium': 'Potassium',
    'sodium': 'Sodium',
    'Hct': 'Hematocrit',
    'Hgb': 'Hemoglobin',
    'PT - INR': 'INR(PT)',
    'MCH': 'MCH',
    'MCHC': 'MCHC',
    'MCV': 'MCV',
    'platelets x 1000': 'Platelet Count',
    'PT': 'PT',
    'PTT': 'PTT',
    'RDW': 'RDW',
    'RBC': 'Red Blood Cells',
    'WBC x 1000': 'White Blood Cells',
    'Heart Rate': 'Heart Rate (bpm)',
    'Non-Invasive BP Diastolic': 'Non Invasive Blood Pressure systolic (mmHg)',
    'Non-Invasive BP Systolic': 'Non Invasive Blood Pressure diastolic (mmHg)',
    'Non-Invasive BP Mean': 'Non Invasive Blood Pressure mean (mmHg)',
    'Respiratory Rate': 'Respiratory Rate (insp/min)',
    'O2 Saturation': 'O2 saturation pulseoxymetry (%)',
    'CI': 'Chloride (serum)',
    'calcium': 'Calcium non-ionized',
    'CPK': 'CK (CPK)',
    'Temperature (F)': 'Temperature Fahrenheit (F)',
    'Pain Score': 'Pain Level',
    'LPM O2': 'O2 Flow (L/min)',
    'O2 L/%': 'Inspired O2 Fraction',
    'ionized calcium': 'Ionized Calcium',
    'albumin': 'Albumin',
    'GCS Total': 'GCS',
    'total bilirubin': 'Total Bilirubin',
    'LDH': 'LDH',
    'ethanol': 'ETOH',
    'Invasive BP Systolic': 'Arterial Blood Pressure systolic (mmHg)',
    'Invasive BP Diastolic': 'Arterial Blood Pressure diastolic (mmHg)',
    'Invasive BP Mean': 'Arterial Blood Pressure mean (mmHg)',
    'serum osmolality':	'Serum Osmolality',
    'troponin - I':	'Troponin-T',
    'uric acid': 'Uric Acid',
    'ammonia': 'Ammonia',
    'CRP': 'C Reactive Protein (CRP)',
    'fibrinogen': 'Fibrinogen',
    'PA Systolic': 'Pulmonary Artery Pressure systolic (mmHg)',
    'PA Diastolic': 'Pulmonary Artery Pressure diastolic (mmHg)',
    'PA Mean': 'Pulmonary Artery Pressure mean (mmHg)',
    'Bedside Glucose': 'Glucose finger stick (range 70-100)',
    'reticulocyte count': 'Reticulocyte Count Automated',
    '-basos': 'Differential-Basos',
    '-eos': 'Differential-Eos',
    '-lymphs': 'Differential-Lymphs',
    '-monos': 'Differential-Monos',
    '-polys': 'Differential-Neuts',
    'haptoglobin': 'Haptoglobin',
    'direct bilirubin': 'Bilirubin Direct',
    'free T4': 'Thyroxine (T4) Free',
    'ESR': 'Sedimentation Rate',
    'CPK-MB INDEX': 'CK-MB',
    'amylase': 'Amylase',
    'PEEP': 'PEEP set (cmH2O)',
    'CVP': 'Central Venous Pressure (mmHg)',
    'unitdischargestatus': 'hospital_expire_flag',
    'LOS': 'los'
}

# Replace the DataFrame and column names mapping
eicu_temp.rename(columns=column_eicu_mapping, inplace=True)



column_mimic_mapping = {
    'Temperature Fahrenheit (°F)': 'Temperature Fahrenheit (F)',
}

# Replace the DataFrame and column names mapping
mimic_temp.rename(columns=column_mimic_mapping, inplace=True)

# Remove "-" from the 'subject_id' column in eicu
eicu_temp['subject_id'] = eicu_temp['subject_id'].str.replace('-', '')

# Convert 'subject_id' in eicu to int64
eicu_temp['subject_id'] = eicu_temp['subject_id'].astype(np.int64)

# Remove duplicate columns
mimic_temp = mimic_temp.loc[:, ~mimic_temp.columns.duplicated()]

# Remove duplicate columns
eicu_temp = eicu_temp.loc[:, ~eicu_temp.columns.duplicated()]

display(mimic_temp)

display(eicu_temp)

# Check if mimic and eicu datasets have the same dtype and header names

# Get the column names and dtypes of mimic_df
mimic_info = mimic_temp.dtypes

# Get the column names and dtypes of eicu_df
eicu_info = eicu_temp.dtypes

# Check if the column names are the same
if all(mimic_info.index == eicu_info.index):
    print("The column names are the same.")
else:
    print("The column names are different.")

print("\n")

# Check if the number of columns is the same
if len(mimic_info) != len(eicu_info):
    print("Number of columns is different between mimic_df and eicu_df.")
else:
    # Iterate over the columns and compare the data type.
    for column_name in mimic_info.index:
        mimic_dtype = mimic_info[column_name]
        eicu_dtype = eicu_info[column_name]
        if mimic_dtype != eicu_dtype:
            print(f"Column '{column_name}' has different data types: mimic_df has '{mimic_dtype}' and eicu_df has '{eicu_dtype}'.")

# Export the merged DataFrame to a CSV file
mimic_temp.to_csv('CSV\\exports\\final\\mimic_mean_final.csv', index=False)

# Export the merged DataFrame to a CSV file
eicu_temp.to_csv('CSV\\exports\\final\\eicu_mean_final.csv', index=False)
